backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9
redis==5.1.1
rq==1.16.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.2
chromadb==0.5.15
pypdf==4.3.1
loguru==0.7.2
httpx==0.27.2
dashscope==1.20.11
python-dotenv==1.0.1
